requirements.txt
# ========= CORE WEB APP =========
streamlit>=1.34.0
extra-streamlit-components==0.1.60

# ========= DATA PROCESSING =========
pandas==2.2.2
numpy==1.26.4

# ========= AUTH & SECURITY =========
bcrypt>=4.0.0
python-dotenv==1.0.1

# ========= HTTP & API =========
requests==2.31.0

# ========= SUPABASE (DB) =========
supabase>=2.0.0
postgrest>=2.0.0

# ========= DATABASE (SQLITE / OPTIONAL PG) =========
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0   # OK trên Windows

# ========= IMAGE PROCESSING =========
Pillow>=10.0.0

# ========= BACKEND (FOR FUTURE API / SERVER) =========
flask>=3.0.0
flask-session>=0.5.0
fastapi==0.110.0
uvicorn[standard]==0.38.0

# ========= AI / ML =========
torch
torchvision
folium
osmnx
google-generativeai
